requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
google-generativeai==0.8.5
schedule==1.2.0
requests==2.31.0
anyio==3.7.1